Add profile editing dialog and clean up DesktopNav; enhance TaskHome logging
