+++ TOKEN_EXPIRATION_UPDATED.md
@@ -1,0 +1,72 @@
+## ✅ Token Expiration Times Updated Successfully!
+
+### 🕐 **New Token Expiration Settings:**
+
+- **Access Token**: **1 Hour** (3600 seconds)
+- **Refresh Token**: **7 Days** (604800 seconds)
+
+### 🔧 **What Was Updated:**
+
+#### 1. **Auth Utility (`src/lib/auth.js`):**
+```javascript
+// Access token: 1 hour (3600 seconds)
+// Refresh token: 7 days (7 * 24 * 60 * 60 = 604800 seconds)
+document.cookie = `access=${accessToken}; path=/; max-age=${60 * 60}; samesite=strict`;
+document.cookie = `refresh=${refreshToken}; path=/; max-age=${7 * 24 * 60 * 60}; samesite=strict`;
+```
+
+#### 2. **Added Token Expiration Utilities:**
+- `isAccessTokenExpired()` - Checks if access token is expired
+- `shouldRefreshToken()` - Checks if token should be refreshed (55 minutes)
+- Token timestamp tracking for accurate expiration checking
+
+#### 3. **Enhanced AxiosProvider:**
+- **Proactive Token Refresh**: Automatically refreshes tokens 5 minutes before expiration
+- **Automatic Retry**: If a request fails with 401, tries to refresh token and retry
+- **Graceful Fallback**: Redirects to login only if refresh fails
+
+#### 4. **Smart Middleware:**
+- Allows access when only refresh token exists (access token expired)
+- Client-side AxiosProvider handles the token refresh seamlessly
+
+### 🔄 **Token Refresh Flow:**
+
+1. **Proactive Refresh** (at 55 minutes):
+   - AxiosProvider automatically refreshes before expiration
+   - Seamless user experience, no interruption
+
+2. **Reactive Refresh** (on 401 error):
+   - If access token expired, automatically tries refresh
+   - Retries the original request with new token
+
+3. **Fallback** (if refresh fails):
+   - Clears all tokens
+   - Redirects to login page
+
+### 📊 **Timeline Example:**
+
+```
+Login: 12:00 PM
+├── Access Token Valid: 12:00 PM - 1:00 PM
+├── Proactive Refresh: 12:55 PM (new access token issued)
+├── Access Token Valid: 12:55 PM - 1:55 PM
+├── Proactive Refresh: 1:50 PM (new access token issued)
+└── Refresh Token Expires: 7 days later at 12:00 PM
+```
+
+### 🎯 **Benefits:**
+
+- ✅ **Secure**: Short-lived access tokens (1 hour)
+- ✅ **User-Friendly**: No unexpected logouts due to proactive refresh
+- ✅ **Automatic**: Handles token refresh transparently
+- ✅ **Robust**: Multiple fallback mechanisms
+
+### 🧪 **Testing Your Setup:**
+
+1. **Login** and check cookies in DevTools (F12 → Application → Cookies)
+2. **Access token** should expire in 1 hour
+3. **Refresh token** should expire in 7 days
+4. **Wait 55+ minutes** to see proactive refresh in action
+5. **API calls** should continue working seamlessly
+
+Your authentication system now properly handles the 1-hour access token and 7-day refresh token lifecycle! 🚀
