+++ package.json
@@ -28,6 +28,7 @@
     "react-dom": "^19.0.0",
     "react-hook-form": "^7.60.0",
     "react-icons": "^5.5.0",
+    "react-loader-spinner": "^6.1.6",
     "react-pdf": "^10.0.1",
     "sweetalert2": "^11.22.2",
     "tailwind-merge": "^3.3.1"
