+++ src/app/chat/page.jsx
@@ -1,10 +1,11 @@
 import ChatHome from "@/components/Chat/Home/ChatHome"
+import ProtectedRoute from "@/components/Auth/ProtectedRoute"
 
 const page = () => {
   return (
     <div className=" lg:px-52 ">
-     <ChatHome/>     
-    </div>
+       <ChatHome/>     
+      </div>
   )
 }
