+++ src/components/HomePage/HeroSection/MentalHealthHero/MentalHealthHero.jsx
@@ -13,10 +13,29 @@ const MentalHealthHero = () => {
           <li>COUPLE COUNSELLING</li>
           <li>FAMILY COUNSELLING</li>
           <li>CHILD COUNSELLING</li>
-           <li>INDIVIDUAL COUNSELLING</li>
-            <li>CHILD COUNSELLING</li>
-              <li>COUPLE COUNSELLING</li>
-         
+          <li>INDIVIDUAL COUNSELLING</li>
+          <li>CHILD COUNSELLING</li>
+          <li>COUPLE COUNSELLING</li>
+          <li>CHILD COUNSELLING</li>
+          <li>FAMILY COUNSELLING</li>
+          <li>INDIVIDUAL COUNSELLING</li>
+          <li>CHILD COUNSELLING</li>
+          <li>COUPLE COUNSELLING</li>
+          <li>FAMILY COUNSELLING</li>
+          <li>CHILD COUNSELLING</li>
+          <li>INDIVIDUAL COUNSELLING</li>
+          <li>CHILD COUNSELLING</li>
+          <li>COUPLE COUNSELLING</li>
+          <li>CHILD COUNSELLING</li>
+          <li>FAMILY COUNSELLING</li>
+          <li>INDIVIDUAL COUNSELLING</li>
+          <li>CHILD COUNSELLING</li>
+          <li>COUPLE COUNSELLING</li>
+          <li>FAMILY COUNSELLING</li>
+          <li>CHILD COUNSELLING</li>
+          <li>INDIVIDUAL COUNSELLING</li>
+          <li>CHILD COUNSELLING</li>
+          <li>COUPLE COUNSELLING</li>
         </div>
       </marquee>
       <hr className="bg-blue-600 h-[1px] border-none mt-4" />
@@ -26,7 +45,9 @@ const MentalHealthHero = () => {
         </h1>
       </div>
       <h1 className="text-center text-white font-bold text-4xl mt-8">
-        Nurture Your Mental <span className="italic font-playfair text-[#0056F6]">Health</span> the Holistic Way
+        Nurture Your Mental{" "}
+        <span className="italic font-playfair text-[#0056F6]">Health</span> the
+        Holistic Way
       </h1>
       {/*Card*/}
       <div className="mt-[60px] lg:flex lg:gap-6 justify-center">
@@ -34,21 +55,27 @@ const MentalHealthHero = () => {
           <h1 className="font-bold text-xl">Step 1</h1>
           <h1 className="font-bold text-xl mt-7 mb-5">Choose Your Concern</h1>
           <p>
-            Choose a concern to begin your journey toward healing and self-discovery.
+            Choose a concern to begin your journey toward healing and
+            self-discovery.
           </p>
         </div>
         <div className=" text-white lg:w-[424px] h-[267px] hover:text-black py-11 px-6 border-1 border-[#76A6FF] hover:bg-[#76A6FF] rounded-xl text-center transition duration-300 mb-2">
           <h1 className="font-bold text-xl">Step 1</h1>
-          <h1 className="font-bold text-xl mt-7 mb-5">Chat or Speak with AI Therapist</h1>
+          <h1 className="font-bold text-xl mt-7 mb-5">
+            Chat or Speak with AI Therapist
+          </h1>
           <p>
-           Choose to chat or speak openly with your AI mental health companion.
+            Choose to chat or speak openly with your AI mental health companion.
           </p>
         </div>
         <div className=" text-white lg:w-[424px] h-[267px] hover:text-black py-11 px-6 border-1  rounded-xl text-center  border-[#76A6FF] hover:bg-[#76A6FF] transition duration-300">
           <h1 className="font-bold text-xl">Step 1</h1>
-          <h1 className="font-bold text-xl mt-7 mb-5">Track Progress & Get Daily Guidance</h1>
+          <h1 className="font-bold text-xl mt-7 mb-5">
+            Track Progress & Get Daily Guidance
+          </h1>
           <p>
-           Monitor your emotional growth over time and receive daily tips to stay on track.
+            Monitor your emotional growth over time and receive daily tips to
+            stay on track.
           </p>
         </div>
       </div>
